renamed pokemon entry interface to remove unnecessary hypen
