+++ src/app/database/database-facade.service.ts
@@ -5,7 +5,7 @@ import { MoveService } from './services/move.service';
 import { LegalityService } from './services/legality.service';
 import { Move } from './models/move.interface';
 import { LegalityEntry } from './models/legality-entry.interface';
-import { PokemonEntry } from './models/pokemon-entry-.interface';
+import { PokemonEntry } from './models/pokemon-entry.interface';
 import { LevelUpMove } from './models/level-up-move.interface';
 import { Database } from './models/database.interface';
