+++ src/app/database/models/pokemon-entry.interface.ts
